% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ds}
\alias{ds}
\title{Creates histogram, boxplot and numeric summary}
\usage{
ds(x)
}
\arguments{
\item{x}{numeric variable}
}
\value{
This is a stupid shit
}
\description{
Creates histogram, boxplot and numeric summary
}
\examples{
ds(33)


}
